requirements.txt
asgiref==3.8.1
coverage==7.6.12
Django
djangorestframework==3.15.2
factory_boy==3.3.3
Faker==36.1.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary<3.0,>=2.9
pytest==8.3.4
pytest-django==4.10.0
sqlparse==0.5.3
tzdata==2025.1
django-debug-toolbar
python-decouple
requests